siraya/requirements_local.txt
supabase
langchain-google-genai
langchain-community
langchain-text-splitters
pypdf
python-dotenv
streamlit>=1.30.0
supabase>=2.0.0
langchain-google-genai>=1.0.0
langchain-core>=0.1.0
pypdf>=4.0.0  # Solo per lo script locale, non serve su Streamlit